+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 * EXOOutboundConnector
   * Fixed issue #821
+* O365OrgCustomizationSetting
+  * Fixes an issue where the resource was not being exported;
+* ODSettings
+  * Removed AD group guid dependency issue # 862
 * SPOTenantSettings
   * Parameter RequireAcceptingAccountMatchInvitedAccount
     is now deprecated (Issue #864)
@@ -23,6 +27,11 @@
   * M365DSCReverse
     * Fixed an issue where passing in the file name and using the GUI for
       extraction did not store the file at the specified location (Issue #810)
+    * Fixed and issue where the -GenerateInfo parameter would always generate
+      a link to the same resource.
+    * Added current version module in the Export file.
+* MISC
+  * Added Authentication Type used to the Telemetry Engine.
 * DEPENDENCIES
   * AzureADPreview Updated to 2.0.2.119
   * DSCParser Updated to 1.3.0.0